backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2023.7.22
click==8.1.8
colorama==0.4.6
distlib==0.3.7
distro==1.9.0
Django==4.2.1
djangorestframework==3.16.0
fastapi==0.115.12
filelock==3.12.3
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.72.0
pipenv==2023.8.28
platformdirs==3.10.0
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.2
uvicorn==0.34.0
virtualenv==20.24.4
